+++ libutilities/ConcurrentCOW.h
@@ -11,6 +11,7 @@ class ConcurrentCOW
 public:
     ConcurrentCOW(){};
     explicit ConcurrentCOW(T&& obj) { reset(std::forward<T>(obj)); };
+    explicit ConcurrentCOW(std::shared_ptr<T>&& obj) { m_obj = std::move(obj); }
     ConcurrentCOW(const ConcurrentCOW& cow) noexcept { m_obj = cow.m_obj; }
     ConcurrentCOW(ConcurrentCOW&& cow) noexcept { m_obj = std::move(cow.m_obj); }
     void operator=(const ConcurrentCOW& cow) noexcept { m_obj = cow.m_obj; }
@@ -24,11 +25,11 @@ public:
     const T& operator*() const noexcept { return get(); }
     const T* operator->() const noexcept { return m_obj.get(); }
 
-    const T& get() const noexcept { return *m_obj; }
+    const T* get() const noexcept { return m_obj.get(); }
 
     long refCount() const noexcept { return m_obj.use_count(); }
 
-    T& mutableGet()
+    T* mutableGet()
     {
         if (refCount() > 1)
         {
@@ -36,7 +37,7 @@ public:
             m_obj = obj;
         }
 
-        return *m_obj;
+        return m_obj.get();
     }
 
 private:
